% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAnalisisDefensivo.R
\name{OptaMAPanalisisdefensivo}
\alias{OptaMAPanalisisdefensivo}
\title{Esta funcion analiza los eventos defensivos y en que zonas se realizan}
\usage{
OptaMAPanalisisdefensivo(df)
}
\arguments{
\item{df}{Un df}
}
\value{
El analisis defensivo a partir de \code{df}
}
\description{
Esta funcion analiza los eventos defensivos y en que zonas se realizan
}
\examples{
OptaMAPanalisisdefensivo(df)
}
