% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPases.R
\name{OptaMAParrowpass}
\alias{OptaMAParrowpass}
\title{Esta funcion crea un mapa indicando con flechas los pases de cada jugador}
\usage{
OptaMAParrowpass(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un número que indica 1 si es local o 0 si es visitante}
}
\value{
El mapa de calor de \code{df} y si es local \code{home}
}
\description{
Esta funcion crea un mapa indicando con flechas los pases de cada jugador
}
\examples{
OptaMAParrowpass(df,1)
OptaMAParrowpass(df,0)
}
