% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCampoFutbol.R
\name{OptaMAPcampofutbol}
\alias{OptaMAPcampofutbol}
\title{Esta funcion dibuja el campo de futbol y viene de la libreria soccermatics}
\usage{
OptaMAPcampofutbol()
}
\description{
Esta funcion dibuja el campo de futbol y viene de la libreria soccermatics
}
\examples{
OptaMAPcampofutbol()
}
