% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCampoFutbol2.R
\name{OptaMAPcampofutbol2}
\alias{OptaMAPcampofutbol2}
\title{Esta funcion crea un mapa con diferentes disparos, muestra su ubicación y calcula su xG.}
\usage{
OptaMAPcampofutbol2()
}
\arguments{
\item{No}{hay parámetros}
}
\value{
El mapa para localizar coordenadas
}
\description{
Esta funcion crea un mapa con diferentes disparos, muestra su ubicación y calcula su xG.
}
\examples{
OptaMAPcampofutbol2()
}
