% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRecuperacionDisparo.R
\name{OptaMAPcontras}
\alias{OptaMAPcontras}
\title{Esta funcion analiza los eventos defensivos de recuperacion de balon que
acaban en disparo en menos de 3 eventos mostrando el tiempo que pasa entre el evento
de tiro y el de recuperacion en segundos}
\usage{
OptaMAPcontras(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un integer 1/0 para saber si es el equipo Local}
}
\value{
El analisis de contras de \code{df} si es el local \code{home}
}
\description{
Esta funcion analiza los eventos defensivos de recuperacion de balon que
acaban en disparo en menos de 3 eventos mostrando el tiempo que pasa entre el evento
de tiro y el de recuperacion en segundos
}
\examples{
OptaMAPcontras <- function(df,1)
}
