% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCorners.R
\name{OptaMAPcorner}
\alias{OptaMAPcorner}
\title{Esta funcion crea un mapa de calor de adonde se lanzan
los corners si son rematados y con que parte del cuerpo}
\usage{
OptaMAPcorner(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un numero que puede ser 1 o 0 en funcion si es local(1) o visitante (0)}
}
\value{
El mapa de \code{df}  y si es local \code{home}
}
\description{
Esta funcion crea un mapa de calor de adonde se lanzan
los corners si son rematados y con que parte del cuerpo
}
\examples{
OptaMAPcorner(df,1)
OptaMAPcorner(df,0)
}
