% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapGrafoPasesAvanzados.R
\name{OptaMAPdfrontpass}
\alias{OptaMAPdfrontpass}
\title{Esta funcion crea el grafo de pases para un equipo durante un partido cuando
los pases son avanzados}
\usage{
OptaMAPdfrontpass(df, home, metros, pases)
}
\arguments{
\item{df}{Un df}

\item{home}{Un numero 1/0 para representar al equipo local o visitante}

\item{metros}{Un integer que representa metros}

\item{pasesmin}{Un numero que indica el numero de pases minimo que tiene en cuenta el grafo}
}
\value{
El grafo de \code{df} cuando el equipo local es \code{home} con más de \code{metros} metros adelante
y un numero de pases \code{pases} minimo
}
\description{
Esta funcion crea el grafo de pases para un equipo durante un partido cuando
los pases son avanzados
}
\examples{
OptaMAPdfrontpass(df,1,3,3)
OptaMAPdfrontpass(df,0,3,3)
}
