% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPasesRadar.R
\name{OptaMAPdirectpass}
\alias{OptaMAPdirectpass}
\title{Esta funcion crea un mapa de radar para cada jugador con las direcciones de pase,
color y numero a partir de la coordenada promedio desde donde pasan}
\usage{
OptaMAPdirectpass(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un numero que indica 1 si es local o 0 si es visitante}
}
\value{
El radar de \code{df} y si es local \code{home}
}
\description{
Esta funcion crea un mapa de radar para cada jugador con las direcciones de pase,
color y numero a partir de la coordenada promedio desde donde pasan
}
\examples{
OptaMAPdirectpass(df,1)
OptaMAPdirectpass(df,0)
}
