% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapHeatFrom.R
\name{OptaMAPheatMapFrom}
\alias{OptaMAPheatMapFrom}
\title{Esta funcion crea un mapa de calor para cada jugador con sus zonas desde donde pasan}
\usage{
OptaMAPheatMapFrom(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un número que indica 1 si es local o 0 si es visitante}
}
\value{
El mapa de calor de \code{df} y si es local \code{home}
}
\description{
Esta funcion crea un mapa de calor para cada jugador con sus zonas desde donde pasan
}
\examples{
OptaMAPheatMapFrom(df,1)
OptaMAPheatMapFrom(df,0)
}
