% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapHeatTo.R
\name{OptaMAPheatMapTo}
\alias{OptaMAPheatMapTo}
\title{Esta funcion crea un mapa de calor para cada jugador con sus zonas adonde pasan}
\usage{
OptaMAPheatMapTo(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un número que indica 1 si es local o 0 si es visitante}
}
\value{
El mapa de calor de \code{df} y si es local \code{home}
}
\description{
Esta funcion crea un mapa de calor para cada jugador con sus zonas adonde pasan
}
\examples{
OptaMAPheatMapTo(df,1)
OptaMAPheatMapTo(df,0)
}
