% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapGrafoPases.R
\name{OptaMAPmatrixpass}
\alias{OptaMAPmatrixpass}
\title{Esta funcion crea el grafo de pases para un equipo durante un partido y
calcula ademas la centralidad de cada jugador y la densidad del grafo}
\usage{
OptaMAPmatrixpass(df, pasesmin)
}
\arguments{
\item{df}{Un df}

\item{pasesmin}{Un numero que indica el numero de pases minimo que tiene en cuenta el grafo}
}
\value{
El grafo de \code{x}  con mas pases de \code{pasesmin}
}
\description{
Esta funcion crea el grafo de pases para un equipo durante un partido y
calcula ademas la centralidad de cada jugador y la densidad del grafo
}
\examples{
OptaMAPmatrixpass(df,1)
OptaMAPmatrixpass(df,10)
}
