% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapGrafoPasesClustering.R
\name{OptaMAPmatrixpasscluster}
\alias{OptaMAPmatrixpasscluster}
\title{Esta funcion crea el grafo de pases para un equipo durante un partido a partir de una
segmentacion kmeans de posiciones. Es decir, si un pase lo hace en la misma zona 2 jugadores
computan como 1, porque lo que se dibuja es el cluster en esa zona de pase independiente quien
lo haya hecho}
\usage{
OptaMAPmatrixpasscluster(df, home, cl)
}
\arguments{
\item{df}{Un df}

\item{home}{Un número de 1 y 0 que indica si es home o away}

\item{cl}{El número de clusters que queremos calcular en la red}
}
\value{
El grafo de \code{df}  del equipo local si \code{home} con un total de \code{cl} clusters
}
\description{
Esta funcion crea el grafo de pases para un equipo durante un partido a partir de una
segmentacion kmeans de posiciones. Es decir, si un pase lo hace en la misma zona 2 jugadores
computan como 1, porque lo que se dibuja es el cluster en esa zona de pase independiente quien
lo haya hecho
}
\examples{
OptaMAPmatrixpasscluster(df,1,20)
OptaMAPmatrixpasscluster(df,1,10)
}
