% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPosesion.R
\name{OptaMAPposesion}
\alias{OptaMAPposesion}
\title{Esta funcion crea un mapa con las posesiones cada 15'}
\usage{
OptaMAPposesion(df)
}
\arguments{
\item{df}{Un df}
}
\value{
El valor de posesiones de \code{df}
}
\description{
Esta funcion crea un mapa con las posesiones cada 15'
}
\examples{
OptaMAPposesion(df)
}
