% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPosesionZonal.R
\name{OptaMAPposesionzonal}
\alias{OptaMAPposesionzonal}
\title{Esta funci칩n crea un gr치fico con el porcentaje de eventos por zonas (18) y poe minutos}
\usage{
OptaMAPposesionzonal(df, minutos)
}
\arguments{
\item{df}{Un df}

\item{minutos}{Un df}
}
\value{
El mapa de posesiones de \code{x} teniendo en cuenta minutos si \code{y}
}
\description{
Esta funci칩n crea un gr치fico con el porcentaje de eventos por zonas (18) y poe minutos
}
\examples{
OptaMAPposesionzonal(df,0)
OptaMAPposesionzonal(df,1)
}
