% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapEventosRadar.R
\name{OptaMAPradarevent}
\alias{OptaMAPradarevent}
\title{Esta funcion crea graficos segun algunas estadisticas de los eventos}
\usage{
OptaMAPradarevent(df, player)
}
\arguments{
\item{df}{Un df}

\item{player}{1/0 según queramos un radar plot por jugador o por posicion
#' @return El mapa de eventos de \code{df} teniendo en cuenta jugador o posicion si \code{player}}
}
\description{
Esta funcion crea graficos segun algunas estadisticas de los eventos
}
\examples{
OptaMAPradarevent(df,0)
OptaMAPradarevent(df,1)
}
