% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapDisparos.R
\name{OptaMAPshoot}
\alias{OptaMAPshoot}
\title{Esta funcion crea un mapa con las zonas de tiro indicando si era gol y el
tamaño según la distancia sobre una porteria y las zonas propuestas por OPTA}
\usage{
OptaMAPshoot(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un numero que puede ser 1 o 0 en funcion si es local(1) o visitante (0)}
}
\value{
El mapa de tiros de \code{df}  y si es local \code{home}
}
\description{
Esta funcion crea un mapa con las zonas de tiro indicando si era gol y el
tamaño según la distancia sobre una porteria y las zonas propuestas por OPTA
}
\examples{
OptaMAPshoot(df,1)
OptaMAPshoot(df,0)
}
