% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapDisparosVista2.R
\name{OptaMAPshoot2}
\alias{OptaMAPshoot2}
\title{Esta funcion crea un mapa con las zonas de tiro indicando si era gol con que parte del
cuerpo se hizo el disparo desde una vista aerea}
\usage{
OptaMAPshoot2(df, home)
}
\arguments{
\item{df}{Un df}

\item{home}{Un numero que puede ser 1 o 0 en funcion si es local(1) o visitante (0)}
}
\value{
El mapa de tiros de \code{df}  y si es local \code{home}
}
\description{
Esta funcion crea un mapa con las zonas de tiro indicando si era gol con que parte del
cuerpo se hizo el disparo desde una vista aerea
}
\examples{
OptaMAPshoot2(df,1)
OptaMAPshoot2(df,0)
}
