% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapDisparosVistaxG.R
\name{OptaMAPshootxG}
\alias{OptaMAPshootxG}
\title{Esta funcion crea un mapa con diferentes disparos, muestra su ubicación y calcula su xG.}
\usage{
OptaMAPshootxG(numeroequipos)
}
\arguments{
\item{numeroequipos}{es un integer que indica el número de equipos para hacer el gráfico}
}
\value{
El mapa de tiros en un partido para los equipos con el xG calculado
}
\description{
Esta funcion crea un mapa con diferentes disparos, muestra su ubicación y calcula su xG.
}
\examples{
OptaMAPshootxG(2)
}
