% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLEventos.R
\name{OptaXMLeventos}
\alias{OptaXMLeventos}
\title{Esta funcion parsea el xml de eventos de opta generando un data frame preparado para usarse}
\usage{
OptaXMLeventos(evento, ...)
}
\arguments{
\item{evento}{Un string con la ruta del fichero}
}
\value{
un df de \code{evento}
}
\description{
Esta funcion parsea el xml de eventos de opta generando un data frame preparado para usarse
}
\examples{
OptaXMLeventos(path)
}
