% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLMatrizPases.R
\name{OptaXMLpassmatrix}
\alias{OptaXMLpassmatrix}
\title{Esta funcion parsea el xml de pases de opta generando un data frame para ser leido con los
datos de los jugadores y sus estadisticas de pase y a que jugadores les pasan.}
\usage{
OptaXMLpassmatrix(x, ...)
}
\arguments{
\item{x}{Un string con la ruta del fichero}
}
\value{
un df de \code{x}
}
\description{
Esta funcion parsea el xml de pases de opta generando un data frame para ser leido con los
datos de los jugadores y sus estadisticas de pase y a que jugadores les pasan.
}
\examples{
OptaXMLpassmatrix(path)
}
