% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLPosesion.R
\name{OptaXMLposesion}
\alias{OptaXMLposesion}
\title{Esta funcion parsea el xml de posesion para la parte territorial cada 15'}
\usage{
OptaXMLposesion(x, ...)
}
\arguments{
\item{x}{Un string con la ruta del fichero}
}
\value{
un df de \code{x}
}
\description{
Esta funcion parsea el xml de posesion para la parte territorial cada 15'
}
\examples{
OptaXMLposesion(path)
}
