% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLEstadisticas.R
\name{OptaXMLstats}
\alias{OptaXMLstats}
\title{Esta funcion parsea el xml de estadisticas}
\usage{
OptaXMLstats(evento, ...)
}
\arguments{
\item{evento}{Un string con la ruta del fichero}
}
\value{
un df de \code{evento}
}
\description{
Esta funcion parsea el xml de estadisticas
}
\examples{
OptaXMLstats(path)
}
